Maket of hw02_task01 is completed

+++ semester2/hw02/hw02_task01/list.h
@@ -9,7 +9,7 @@ public:
     virtual void insert(int position, ListItem item) = 0;
     virtual void remove(int position) = 0;
     virtual int find(ListItem item) = 0;
-    virtual size_t length() = 0;
+    virtual int length() = 0;
     virtual ListItem & operator [](int position) = 0;
 };
